requirements.txt
fastapi==0.70.0
numpy ==1.24.3
Pillow ==9.5.0